fix(jetcobot): PostureData angle type to int

+++ TechnicalResearch/jetcobot/app/model/model.py
@@ -17,5 +17,5 @@ class PosturesData(BaseModel):
     j4: float
     j5: float
     j6: float
-    angle: float
+    angle: int
     gap: int
